Moved document type back to resident model directory
